+++ ll/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Changed
+- Updated `embedded-hal` from `1.0.0-rc.11` to `1.0.0-rc.2`.
+
 ## [0.12.0] - 2023-10-02
 ### Changed
 - Updated `embedded-hal` from `1.0.0-alpha.11` to `1.0.0-rc.1`.