+++ ll/CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Updated `embedded-hal` from `1.0.0-rc.11` to `1.0.0-rc.3`.
 
+## Fixed
+- Added a second chip-select deassert to meet minimum CS timings with `eh0`.
+
 ## [0.12.0] - 2023-10-02
 ### Changed
 - Updated `embedded-hal` from `1.0.0-alpha.11` to `1.0.0-rc.1`.