+++ ll/CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Changed
-- Updated `embedded-hal` from `1.0.0-rc.11` to `1.0.0-rc.3`.
+- Updated `embedded-hal` from `1.0.0-rc.11` to `1`.
 - Renamed the `eha0a` feature to `eha1`.
 
 ### Fixed