+++ ll/CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.13.0] - 2024-06-09
 ### Changed
 - Updated `embedded-hal` from `1.0.0-rc.11` to `1`.
 - Renamed the `eha0a` feature to `eha1`.
@@ -192,7 +192,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.1.0] - 2021-01-02
 - Initial release
 
-[Unreleased]: https://github.com/newAM/w5500-rs/compare/ll%2Fv0.12.0...HEAD
+[Unreleased]: https://github.com/newAM/w5500-rs/compare/ll%2Fv0.13.0...HEAD
+[0.13.0]: https://github.com/newAM/w5500-rs/compare/ll%2Fv0.12.0...ll%2Fv0.13.0
 [0.12.0]: https://github.com/newAM/w5500-rs/compare/ll%2Fv0.11.0...ll%2Fv0.12.0
 [0.11.0]: https://github.com/newAM/w5500-rs/compare/ll%2Fv0.10.4...ll%2Fv0.11.0
 [0.10.4]: https://github.com/newAM/w5500-rs/compare/ll%2Fv0.10.3...ll%2Fv0.10.4